+++ src/api/egl/mod.rs
@@ -12,6 +12,7 @@ use ReleaseBehavior;
 use Robustness;
 use Api;
 
+use std::cell::Cell;
 use std::ffi::{CStr, CString};
 use std::os::raw::{c_void, c_int};
 use std::{mem, ptr};
@@ -38,9 +39,10 @@ pub struct Context {
     egl: ffi::egl::Egl,
     display: ffi::egl::types::EGLDisplay,
     context: ffi::egl::types::EGLContext,
-    surface: ffi::egl::types::EGLSurface,
+    surface: Cell<ffi::egl::types::EGLSurface>,
     api: Api,
     pixel_format: PixelFormat,
+    config_id: ffi::egl::types::EGLConfig,
 }
 
 #[cfg(target_os = "android")]
@@ -267,8 +269,42 @@ impl Context {
         })
     }
 
+    // Handle Android Life Cycle.
+    // Android has started the activity or sent it to foreground.
+    // Create a new surface and attach it to the recreated ANativeWindow.
+    // Restore the EGLContext.
+    #[cfg(target_os = "android")]
+    pub unsafe fn on_surface_created(&self, native_window: ffi::EGLNativeWindowType) {
+        self.surface.set(self.egl.CreateWindowSurface(self.display, self.config_id, native_window, ptr::null()));
+        if self.surface.get().is_null() {
+            panic!("on_surface_created: eglCreateWindowSurface failed")
+        }
+        let ret = self.egl.MakeCurrent(self.display, self.surface.get(), self.surface.get(), self.context);
+        if ret == 0 {
+            panic!("on_surface_created: eglMakeCurrent failed");
+        }
+    }
+
+    // Handle Android Life Cycle.
+    // Android has stopped the activity or sent it to background.
+    // Release the surface attached to the destroyed ANativeWindow.
+    // The EGLContext is not destroyed so it can be restored later.
+    #[cfg(target_os = "android")]
+    pub unsafe fn on_surface_destroyed(&self) {
+        let ret = self.egl.MakeCurrent(self.display, ffi::egl::NO_SURFACE, ffi::egl::NO_SURFACE, ffi::egl::NO_CONTEXT);
+        if ret == 0 {
+            panic!("on_surface_destroyed: eglMakeCurrent failed");
+        }
+
+        self.egl.DestroySurface(self.display, self.surface.get());
+        self.surface.set(ffi::egl::NO_SURFACE);
+    }
+
     pub unsafe fn make_current(&self) -> Result<(), ContextError> {
-        let ret = self.egl.MakeCurrent(self.display, self.surface, self.surface, self.context);
+        if self.surface.get() == ffi::egl::NO_SURFACE {
+            return Err(ContextError::ContextLost);
+        }
+        let ret = self.egl.MakeCurrent(self.display, self.surface.get(), self.surface.get(), self.context);
 
         if ret == 0 {
             match self.egl.GetError() as u32 {
@@ -283,6 +319,9 @@ impl Context {
 
     #[inline]
     pub fn is_current(&self) -> bool {
+        if self.surface.get() == ffi::egl::NO_SURFACE {
+            return false;
+        }
         unsafe { self.egl.GetCurrentContext() == self.context }
     }
 
@@ -296,13 +335,22 @@ impl Context {
 
     #[inline]
     pub fn swap_buffers(&self) -> Result<(), ContextError> {
+        if self.surface.get() == ffi::egl::NO_SURFACE {
+            return Err(ContextError::ContextLost);
+        }
         let ret = unsafe {
-            self.egl.SwapBuffers(self.display, self.surface)
+            self.egl.SwapBuffers(self.display, self.surface.get())
         };
 
         if ret == 0 {
             match unsafe { self.egl.GetError() } as u32 {
                 ffi::egl::CONTEXT_LOST => return Err(ContextError::ContextLost),
+                // 'on_surface_destroyed' Android event can arrive with some delay because multithreading communication.
+                // Because of that, swap_buffers can be called before processing 'on_surface_destroyed' event, with the
+                // native window surface already destroyed. EGL generates a BAD_SURFACE error in this situation.
+                // We treat it as CONTEXT_LOST error to avoid a panic in servo.
+                // The surface will be restored when the app comes to foreground.
+                #[cfg(target_os = "android")] ffi::egl::BAD_SURFACE => return Err(ContextError::ContextLost),
                 err => panic!("eglSwapBuffers failed (eglGetError returned 0x{:x})", err)
             }
 
@@ -336,7 +384,9 @@ impl Drop for Context {
             // we don't call MakeCurrent(0, 0) because we are not sure that the context
             // is still the current one
             self.egl.DestroyContext(self.display, self.context);
-            self.egl.DestroySurface(self.display, self.surface);
+            if self.surface.get() != ffi::egl::NO_SURFACE {
+                self.egl.DestroySurface(self.display, self.surface.get());
+            }
             self.egl.Terminate(self.display);
         }
     }
@@ -451,9 +501,10 @@ impl<'a> ContextPrototype<'a> {
             egl: self.egl,
             display: self.display,
             context: context,
-            surface: surface,
+            surface: Cell::new(surface),
             api: self.api,
             pixel_format: self.pixel_format,
+            config_id: self.config_id,
         })
     }
 }