+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Android life cycle improvements
+
 # Version 0.12.1 (2018-02-05)
 
 - Add support for winapi 0.3 ([#975](https://github.com/tomaka/glutin/pull/975))